undo rename of package folder
