+++ .github/workflows/publish-npm-github-packages.yml
@@ -32,8 +32,8 @@ jobs:
           node-version: "24.x"
           registry-url: "https://registry.npmjs.org"
       - run: yarn install
-      - run: yarn --cwd module build
-      - run: yarn --cwd module publish
+      - run: yarn --cwd package build
+      - run: yarn --cwd package publish
         env:
           NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
 
@@ -50,7 +50,7 @@ jobs:
           node-version: "24.x"
           registry-url: https://npm.pkg.github.com/
       - run: yarn install
-      - run: yarn --cwd module build
-      - run: yarn --cwd module publish
+      - run: yarn --cwd package build
+      - run: yarn --cwd package publish
         env:
           NODE_AUTH_TOKEN: ${{secrets.NPM_PUBLISH_TOKEN}}
