+++ .gitignore
@@ -1,4 +1,4 @@
-module/coverage/
+package/coverage/
 dist/
 node_modules/
 yarn-error.log
