+++ app/src/ui/LensTable/Table/Row/Aperture.tsx
@@ -1,5 +1,5 @@
 import { FormControl, InputLabel, MenuItem, Select, SelectChangeEvent } from '@mui/material'
-import { apertureMap } from '../../../../../../module/dist'
+import { apertureMap } from '../../../../../../package/dist'
 import useDoFStore from '../../../../store'
 import useIsMobile from '../../../../utilities/useIsMobile'
