+++ package.json
@@ -3,17 +3,17 @@
     "version": "3.1.0",
     "private": true,
     "workspaces": [
-        "module",
+        "package",
         "app"
     ],
     "scripts": {
-        "dev": "yarn --cwd module tsc && yarn --cwd app dev",
+        "dev": "yarn --cwd package tsc && yarn --cwd app dev",
         "start": "yarn --cwd app start",
         "start:ci": "yarn build && yarn --cwd app start",
-        "build": "yarn --cwd module lint && yarn --cwd module tsc && yarn --cwd app build",
-        "deploy": "yarn --cwd module lint && yarn --cwd module tsc && IS_DEPLOYMENT=true yarn --cwd app build",
+        "build": "yarn --cwd package lint && yarn --cwd package tsc && yarn --cwd app build",
+        "deploy": "yarn --cwd package lint && yarn --cwd package tsc && IS_DEPLOYMENT=true yarn --cwd app build",
         "clean": "git clean -fdx",
-        "cover": "yarn --cwd module test --collectCoverage",
+        "cover": "yarn --cwd package test --collectCoverage",
         "lint": "yarn workspaces run lint",
         "preversion": "yarn workspaces run preversion",
         "postversion": "git push && git push --tags",
