+++ package/.ts-prunerc
