+++ package/src/calculators/calculateAperture.test.ts
