+++ package/src/calculators/calculateAperture.ts
