+++ package/src/calculators/calculateCropFactor.ts
