+++ package/src/calculators/calculateFocalLength.ts
