+++ package/src/data/preciseApertureMap.ts
