+++ package/src/index.ts
