+++ package/src/utilities/combineSettings.ts
